# [Mono.Android] Faster native members registration