Fixes https://github.com/xamarin/xamarin-android/issues/1898

Measured results for XA template start-up (x86 emulator):

| RegisterNativeMembers | Before | After | After with new marshal methods |
|-----------------------|-------:|------:|-------------------------------:|
| Debug                 |  169ms |  33ms |                            N/A |
| Release               |   56ms |  18ms |                            8ms |

For more comparison, RegisterNativeMembers in the 15.8/Release takes
27ms.

The `JniTypeManager:TryRegisterNativeMembers` was quite slow, it used
reflection to find the registration method at runtime.

The new implementation finds the method during build in the Linker and
emits code to prefill a dictionary and to call the registration
methods at runtime.

That removes most of the reflection code calls, most importantly the
slow ones.

Possible future improvements:

* find better way to get the index for a given type. replace
  dictionary with perfect hash function to map type names to indexes
  or add our index to the `typemap.mj` table.

* find out a way to avoid using `MethodInfo::CreateDelegate (Type)`
  I already tried that, but wasn't able to get rid of the
  `MethodAccessException`, so had to get back to `CreateDelegate` to
  avoid that. Still think it might be possible, just need more time to
  look into it. At this point we are quite fast already though and
  there are probably more important things to optimize now. (we spend
  about 1ms in these `CreateDelegate` calls)

An example of the linker updated `Prefill` method:
```
private static void Prefill ()
{
	MagicRegistrationMap.typesMap ["xatemplateaot.MainActivity"] = 0;
	MagicRegistrationMap.typesMap ["Android.Runtime.InputStreamAdapter"] = 1;
	MagicRegistrationMap.typesMap ["Android.Runtime.JavaObject"] = 2;
	MagicRegistrationMap.typesMap ["Android.Runtime.OutputStreamAdapter"] = 3;
        ...
	MagicRegistrationMap.typesMap ["Java.Interop.TypeManager+JavaTypeManager"] = 94;
}
```

An example of the linker updated `CallRegisterMethodByIndex` method (the IL -> C# disassembler is a bit misleading at the end, where we return *false* where registration method hasn't been called):
```
private static bool CallRegisterMethodByIndex (JniNativeMethodRegistrationArguments arguments, int typeIdx)
{
	object handle;
	switch (typeIdx) {
	case 0:
		handle = typeof(MainActivity.__<$>_jni_marshal_methods).TypeHandle;
		break;
	case 1:
		handle = typeof(InputStreamAdapter.__<$>_jni_marshal_methods).TypeHandle;
		break;
	case 2:
		handle = typeof(Android.Runtime.JavaObject.__<$>_jni_marshal_methods).TypeHandle;
		break;
	case 3:
		handle = typeof(OutputStreamAdapter.__<$>_jni_marshal_methods).TypeHandle;
		break;
        ...
	case 94:
		handle = typeof(Java.Interop.TypeManager.JavaTypeManager.__<$>_jni_marshal_methods).TypeHandle;
		break;
	}
	((Action<JniNativeMethodRegistrationArguments>)Type.GetTypeFromHandle ((RuntimeTypeHandle)handle).GetMethod ("__RegisterNativeMembers").CreateDelegate (typeof(Action<JniNativeMethodRegistrationArguments>))) (arguments);
	return true;
	IL_0538:
	return false;
}
```