@jonpryor 2018-07-27
Is this Ready For Use, even experimentally?

Can we update e.g. `src/Mono.Android/Test` to make use of this, and see if the changes are visible in our Jenkins Plots?