@radekdoulik 2018-07-31
> Is this Ready For Use, even experimentally?

Yes, definitely. The Before/After column values are for XA template app without usage of the new marshal methods.

The case of using the new marshal methods is not enabled yet, as the build process changes are not committed yet. There might still be some adjustments and fixes in that area.

BTW, today I got a nice idea of doing partial (profiled) generation of the marshal methods (https://github.com/xamarin/xamarin-android/issues/2021). That would allow us to have faster startup without much added size. That should work nicely even in Debug configuration. We can include types used during normal XA template and XA/XForms template startup in `Mono.Android.dll`.

I also put down an observation about the new marshaling, where the last part which avoids us getting even better times with the new marshaling are the constructors. That still uses the old path and thus pulls in System.Reflection.Emit code. I think once we fix that, we might even get somewhere around 4ms time for `RegisterNativeMembers` during the XA template startup. (https://github.com/xamarin/xamarin-android/issues/2022)