@jonpryor 2018-08-01
Given that `Prefill()` generates statically known data, does this really need to be a `Dictionary<string, int>`?

How does the current dictionary compare to a static *sorted* array, a'la:

```csharp
static string[] TypesMap = new[]{
    "xatemplateaot.MainActivity",
    // ...
    "Java.Interop.TypeManager+JavaTypeManager",
};
```

This could be searched through in O(log n) time.  This is certainly theoretically slower than a theoretical O(1) for the Dictionary, but the Dictionary still needs to be *constructed*, which isn't "free".  (Array creation isn't free either, for that matter, but it almost certainly involves allocating less memory than the dictionary.)

For example, consider this (really stupid) [`dict-vs-array-creation.cs`](https://gist.github.com/jonpryor/9bf3ee0c36cd96cb19468ae29f2cc527) program:

```bash
$ csc dict-vs-array-creation.cs
$ mono dict-vs-array-creation.exe
create array: 00:00:01.7723561
 create dict: 00:00:11.1699148
lookup array: 00:00:06.1373072
 lookup dict: 00:00:00.2376908
```

An array with 1000 strings it's allocated in ~1/6 the time the as a Dictionary with 1000 entries.

Looking up a random value via binary search on that array takes ~30x longer than the Dictionary.  (Behold the difference between O(log n) vs. O(1)!)

In the context of Xamarin.Android, which of these is "better"?  Faster allocation should permit faster startup -- presumably? -- but there is a point where the type lookups will slow things down.  Will that slowdown *also* impact startup, or will it be a net gain (faster startup at the cost of more overhead as time goes on)?

Meanwhile, if we make *one* change to the creation metric and provide a size to the dictionary:

```csharp
    var d = new Dictionary<string, int>(1000) {
```

That *hugely* improves things:

```bash
create array: 00:00:01.9271051
 create dict: 00:00:04.9936653
lookup array: 00:00:06.1666141
 lookup dict: 00:00:00.2579320
```

Dictionary construction time is *halved*, meaning the array construction os now only ~1/3, not ~1/6, of the Dictionary time.