@radekdoulik 2018-08-01
> Given that Prefill() generates statically known data, does this really need to be a Dictionary<string, int>?

It doesn't and as I wrote in the commit message, there are better ways to do it, as the perfect hashing for example. The current version with the dictionary works reasonably well, the 'Prefill' takes 1ms in the XA template I measured.

Moving the dictionary creation to `Prefill` is one of the adjustments I have in mind. I wanted to get the whole patch out as I spent a lot time on it and wanted things to move along.

I can update the dictionary initialization in the new PR. Calling the constructor with the size is nice improvement, good idea!

Meanwhile I would like to fix cecil (or linker/jnimarshalmethod-gen?) problem, where we are breaking virtual classes and getting `System.TypeLoadException` on linked assemblies.